Added final code

+++ src/main/java/customevents/model/LogEvents.java
@@ -80,7 +80,7 @@ public class LogEvents {
 	/**
 	 * @return the timestamp
 	 */
-	public float getTimestamp() {
+	public long getTimestamp() {
 		return timestamp;
 	}
